# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Revenue Recovery AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
  request_id?: string;
}

export type RecoveryStatus = typeof RecoveryStatus[keyof typeof RecoveryStatus];


export const RecoveryStatus = {
  active: 'active',
  recovered: 'recovered',
  failed: 'failed',
  paused: 'paused',
  exhausted: 'exhausted',
} as const;

export type RecoveryStrategy = typeof RecoveryStrategy[keyof typeof RecoveryStrategy];


export const RecoveryStrategy = {
  smart_retry: 'smart_retry',
  email: 'email',
  whatsapp: 'whatsapp',
  sms: 'sms',
  support: 'support',
} as const;

export interface Recovery {
  id: string;
  payment_id: string;
  customer_id?: string | null;
  amount: number;
  currency: string;
  root_cause?: string | null;
  status: RecoveryStatus;
  strategies: RecoveryStrategy[];
  selected_strategy?: RecoveryStrategy | null;
  attempts_used: number;
  max_attempts: number;
  recovery_amount?: number | null;
  created_at: string;
  updated_at: string;
}

export interface RecoveryCreate {
  payment_id: string;
  customer_id?: string;
  /** @minimum 1 */
  amount: number;
  currency?: string;
  root_cause?: string;
  strategies?: RecoveryStrategy[];
}

export interface Pagination {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
}

export interface RecoveryListResponse {
  data: Recovery[];
  pagination: Pagination;
}

export interface TimelineEvent {
  id: string;
  title: string;
  description: string;
  status: string;
  timestamp: string;
}

export type AuditLogDetails = { [key: string]: unknown };

export interface AuditLog {
  id: string;
  action: string;
  actor: string;
  details: AuditLogDetails;
  created_at: string;
}

export type RecoveryDetail = Recovery & {
  timeline: TimelineEvent[];
  audit_logs: AuditLog[];
};

export interface TrendPoint {
  date: string;
  failed: number;
  recovered: number;
  recovered_amount: number;
}

export interface DashboardResponse {
  total_failed: number;
  total_recovered: number;
  recovery_attempts: number;
  recovery_rate: number;
  revenue_at_risk: number;
  revenue_recovered: number;
  active_recoveries: number;
  trend: TrendPoint[];
}

export type AnalyticsResponseStrategyPerformance = {[key: string]: number};

export type AnalyticsResponseCostBreakdown = {[key: string]: number};

export type AnalyticsResponseRevenueByStrategy = {[key: string]: number};

export interface TopCustomer {
  customer_id: string;
  recovered_amount: number;
  recovery_count: number;
}

export interface AnalyticsResponse {
  recovery_rate: number;
  revenue_impact: number;
  total_attempts: number;
  successful_attempts: number;
  strategy_performance: AnalyticsResponseStrategyPerformance;
  cost_breakdown: AnalyticsResponseCostBreakdown;
  revenue_by_strategy: AnalyticsResponseRevenueByStrategy;
  top_customers: TopCustomer[];
}

export type SettingsAccount = { [key: string]: unknown };

export type SettingsRecovery = { [key: string]: unknown };

export type SettingsIntegrations = { [key: string]: unknown };

export interface Settings {
  account: SettingsAccount;
  recovery: SettingsRecovery;
  integrations: SettingsIntegrations;
}

export interface SettingsUpdate { [key: string]: unknown }

export type RazorpayWebhookPayload = { [key: string]: unknown };

export interface RazorpayWebhook {
  event: string;
  payload: RazorpayWebhookPayload;
  contains?: string[];
  account_id?: string | null;
}

export interface WebhookResponse {
  received: boolean;
  event_id: string;
  duplicate: boolean;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Server error
 */
export type ServerErrorResponse = ErrorResponse;

export type DaysParameter = number;

export type PageParameter = number;

export type PerPageParameter = number;

export type GetDashboardParams = {
/**
 * @minimum 1
 * @maximum 365
 */
days?: DaysParameter;
};

export type GetDashboardMetricsParams = {
/**
 * @minimum 1
 * @maximum 365
 */
days?: DaysParameter;
};

export type ListRecoveriesParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 100
 */
per_page?: PerPageParameter;
status?: RecoveryStatus;
search?: string;
};

export type GetAnalyticsParams = {
start_date?: string;
end_date?: string;
};