# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Revenue Recovery AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get recovery analytics
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
...
  responses:
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        request_id:
          type: string
      required: [error, message]
...
    AnalyticsResponse:
      type: object
      properties:
        recovery_rate:
          type: number
        revenue_impact:
          type: number
        total_attempts:
          type: number
        successful_attempts:
          type: number
        strategy_performance:
          type: object
          additionalProperties:
            type: number
        cost_breakdown:
          type: object
          additionalProperties:
            type: number
        revenue_by_strategy:
          type: object
          additionalProperties:
            type: number
        top_customers:
          type: array
          items:
            $ref: "#/components/schemas/TopCustomer"
      required: [recovery_rate, revenue_impact, total_attempts, successful_attempts, strategy_performance, cost_breakdown, revenue_by_strategy, top_customers]
    TopCustomer:
      type: object
      properties:
        customer_id:
          type: string
        recovered_amount:
          type: number
        recovery_count:
          type: number
      required: [customer_id, recovered_amount, recovery_count]
...